# Improve avg/sum Aggregator performance