--- a/datafusion/expr/src/aggregate_function.rs
+++ b/datafusion/expr/src/aggregate_function.rs
@@ -161,6 +161,19 @@ pub fn return_type(
     }
 }
 
+/// Returns the internal sum datatype of the avg aggregate function.
+pub fn sum_type_of_avg(input_expr_types: &[DataType]) -> Result<DataType> {
+    // Note that this function *must* return the same type that the respective physical expression returns
+    // or the execution panics.
+    let fun = AggregateFunction::Avg;
+    let coerced_data_types = crate::type_coercion::aggregates::coerce_types(
+        &fun,
+        input_expr_types,
+        &signature(&fun),
+    )?;
+    avg_sum_type(&coerced_data_types[0])
+}
+
 /// the signatures supported by the function `fun`.
 pub fn signature(fun: &AggregateFunction) -> Signature {
     // note: the physical expression must accept the type returned by this function or the execution panics.