@Dandandan 2023-04-04
@mingmwang query 18 also is affected by this (although should be less of a difference than query 17)