@andygrove 2023-04-04
Here are the results on my desktop (24 core) for queries 17 and 18.

## This PR

```
Executing query 17 from /home/andy/git/sql-benchmarks/sqlbench-h/queries/sf=10//q17.sql
Query 17 executed in: 25.076557605s
Query 17 executed in: 14.468542149s
Query 17 executed in: 13.238961128s
Executing query 18 from /home/andy/git/sql-benchmarks/sqlbench-h/queries/sf=10//q18.sql
Query 18 executed in: 8.197590942s
Query 18 executed in: 7.883707736s
Query 18 executed in: 7.936216688s
```

## Main branch

```
Executing query 17 from /home/andy/git/sql-benchmarks/sqlbench-h/queries/sf=10//q17.sql
Query 17 executed in: 20.10441009s
Query 17 executed in: 14.731355792s
Query 17 executed in: 13.869329179s
Executing query 18 from /home/andy/git/sql-benchmarks/sqlbench-h/queries/sf=10//q18.sql
Query 18 executed in: 8.900509355s
Query 18 executed in: 8.695394752s
Query 18 executed in: 8.541570444s
Executing query 19 from /home/andy/git
```