@Dandandan 2023-04-04
When running in memory (SF=1)

```
cargo run --release --bin tpch benchmark datafusion --path ./data/ --format parquet --partitions 16 -q 17 --iterations 10 -m
```

main
==
```
Query 17 iteration 0 took 725.6 ms and returned 1 rows
Query 17 iteration 1 took 675.3 ms and returned 1 rows
Query 17 iteration 2 took 635.1 ms and returned 1 rows
Query 17 iteration 3 took 702.1 ms and returned 1 rows
Query 17 iteration 4 took 624.7 ms and returned 1 rows
Query 17 iteration 5 took 692.2 ms and returned 1 rows
Query 17 iteration 6 took 681.0 ms and returned 1 rows
Query 17 iteration 7 took 668.9 ms and returned 1 rows
Query 17 iteration 8 took 675.6 ms and returned 1 rows
Query 17 iteration 9 took 660.6 ms and returned 1 rows
Query 17 avg time: 674.12 ms
```

```
Query 18 iteration 0 took 377.6 ms and returned 57 rows
Query 18 iteration 1 took 361.1 ms and returned 57 rows
Query 18 iteration 2 took 353.7 ms and returned 57 rows
Query 18 iteration 3 took 362.2 ms and returned 57 rows
Query 18 iteration 4 took 359.8 ms and returned 57 rows
Query 18 iteration 5 took 356.8 ms and returned 57 rows
Query 18 iteration 6 took 350.8 ms and returned 57 rows
Query 18 iteration 7 took 360.6 ms and returned 57 rows
Query 18 iteration 8 took 351.6 ms and returned 57 rows
Query 18 iteration 9 took 352.2 ms and returned 57 rows
Query 18 avg time: 358.65 ms
```

PR
==
```
Query 17 iteration 0 took 640.5 ms and returned 1 rows
Query 17 iteration 1 took 649.0 ms and returned 1 rows
Query 17 iteration 2 took 597.3 ms and returned 1 rows
Query 17 iteration 3 took 639.3 ms and returned 1 rows
Query 17 iteration 4 took 645.2 ms and returned 1 rows
Query 17 iteration 5 took 653.8 ms and returned 1 rows
Query 17 iteration 6 took 628.6 ms and returned 1 rows
Query 17 iteration 7 took 643.7 ms and returned 1 rows
Query 17 iteration 8 took 600.0 ms and returned 1 rows
Query 17 iteration 9 took 591.8 ms and returned 1 rows
Query 17 avg time: 628.91 ms
```

```
Query 18 iteration 0 took 355.8 ms and returned 57 rows
Query 18 iteration 1 took 333.5 ms and returned 57 rows
Query 18 iteration 2 took 317.5 ms and returned 57 rows
Query 18 iteration 3 took 327.6 ms and returned 57 rows
Query 18 iteration 4 took 318.2 ms and returned 57 rows
Query 18 iteration 5 took 324.2 ms and returned 57 rows
Query 18 iteration 6 took 315.0 ms and returned 57 rows
Query 18 iteration 7 took 323.1 ms and returned 57 rows
Query 18 iteration 8 took 314.8 ms and returned 57 rows
Query 18 iteration 9 took 322.0 ms and returned 57 rows
Query 18 avg time: 325.17 ms
```