@Dandandan 2023-04-04
Although PR shows an improvement, ~15% of the CPU samples still end up in `cast`.
![flamegraph](https://user-images.githubusercontent.com/163737/229907153-fc92cbc4-0d4c-40f9-b1e8-aafec75df659.svg)