@mingmwang 2023-04-05
I run the test with sf=1 and partition = 1.  I will do more test with sf= 10 tomorrow.