@yahoNanJing 2023-04-06
@Dandandan, how did you generate the flame graph? At my side, there're many unknowns. Could you teach me how to avoid them.
![perf-0](https://user-images.githubusercontent.com/90197956/230261813-ccb5b7a3-173e-44c1-b89a-cc82503a436b.svg)

`
cargo run --release --bin tpch -- benchmark datafusion --iterations 10 --path ./data-parquet --format parquet --partitions 1 --query 17
`

`sudo perf record -F 99 -g -p 917304`

`sudo perf script | ./stackcollapse-perf.pl | ./flamegraph.pl > perf.svg`