@yahoNanJing 2023-04-06
Hi @Dandandan, for the latest code, the **_cast_** is almost avoided. The related flame graphs are as follows:
![perf-5866](https://user-images.githubusercontent.com/90197956/230296399-78aec1e7-0d0f-448a-9454-63861cb18426.svg)
![perf-main](https://user-images.githubusercontent.com/90197956/230296407-662baf44-db00-44cc-b6ef-c382f09548eb.svg)

With single partition benchmark, the latency is reduced from 11s to 9s for q17.